refactor: Adjusting according to suggestions
